+++ firebase-firestore/src/main/java/com/google/firebase/firestore/UserDataConverter.java
@@ -70,17 +70,17 @@ public final class UserDataConverter {
   }
 
   /** Parse document data from a non-merge set() call. */
-  public ParsedSetData parseSetData(Map<String, Object> input) {
+  public ParsedSetData parseSetData(Object input) {
     ParseAccumulator accumulator = new ParseAccumulator(UserData.Source.Set);
-    FieldValue updateData = parseData(input, accumulator.rootContext());
+    FieldValue updateData = convertAndParseData(input, accumulator.rootContext());
 
     return accumulator.toSetData((ObjectValue) updateData);
   }
 
   /** Parse document data from a set() call with SetOptions.merge() set. */
-  public ParsedSetData parseMergeData(Map<String, Object> input, @Nullable FieldMask fieldMask) {
+  public ParsedSetData parseMergeData(Object input, @Nullable FieldMask fieldMask) {
     ParseAccumulator accumulator = new ParseAccumulator(UserData.Source.MergeSet);
-    ObjectValue updateData = (ObjectValue) parseData(input, accumulator.rootContext());
+    ObjectValue updateData = (ObjectValue) convertAndParseData(input, accumulator.rootContext());
 
     if (fieldMask != null) {
       // Verify that all elements specified in the field mask are part of the parsed context.
@@ -118,7 +118,8 @@ public final class UserDataConverter {
         // Add it to the field mask, but don't add anything to updateData.
         context.addToFieldMask(fieldPath);
       } else {
-        @Nullable FieldValue parsedValue = parseData(fieldValue, context.childContext(fieldPath));
+        @Nullable
+        FieldValue parsedValue = convertAndParseData(fieldValue, context.childContext(fieldPath));
         if (parsedValue != null) {
           context.addToFieldMask(fieldPath);
           updateData = updateData.set(fieldPath, parsedValue);
@@ -168,7 +169,7 @@ public final class UserDataConverter {
         // Add it to the field mask, but don't add anything to updateData.
         context.addToFieldMask(parsedField);
       } else {
-        FieldValue parsedValue = parseData(fieldValue, context.childContext(parsedField));
+        FieldValue parsedValue = convertAndParseData(fieldValue, context.childContext(parsedField));
         if (parsedValue != null) {
           context.addToFieldMask(parsedField);
           updateData = updateData.set(parsedField, parsedValue);
@@ -183,7 +184,7 @@ public final class UserDataConverter {
   public FieldValue parseQueryValue(Object input) {
     ParseAccumulator accumulator = new ParseAccumulator(UserData.Source.Argument);
 
-    @Nullable FieldValue parsed = parseData(input, accumulator.rootContext());
+    @Nullable FieldValue parsed = convertAndParseData(input, accumulator.rootContext());
     hardAssert(parsed != null, "Parsed data should not be null.");
     hardAssert(
         accumulator.getFieldTransforms().isEmpty(),
@@ -191,32 +192,31 @@ public final class UserDataConverter {
     return parsed;
   }
 
-  /**
-   * Converts a POJO into a Map, throwing appropriate errors if it wasn't actually a proper POJO.
-   */
-  public Map<String, Object> convertPOJO(Object pojo) {
-    checkNotNull(pojo, "Provided data must not be null.");
-    String reason =
+  /** Converts a POJO to native types and then parses it into model types. */
+  private FieldValue convertAndParseData(Object input, ParseContext context) {
+    UserData.Source source = context.getDataSource();
+    String badDocReason =
         "Invalid data. Data must be a Map<String, Object> or a suitable POJO object, but it was ";
 
     // Check Array before calling CustomClassMapper since it'll give you a confusing message
-    // to use List instead, which also won't work.
-    if (pojo.getClass().isArray()) {
-      throw new IllegalArgumentException(reason + "an array");
+    // to use List instead, which also won't work in a set().
+    if (source == UserData.Source.Set && input.getClass().isArray()) {
+      throw new IllegalArgumentException(badDocReason + "an array");
     }
 
-    Object converted = CustomClassMapper.convertToPlainJavaTypes(pojo);
-    if (!(converted instanceof Map)) {
-      throw new IllegalArgumentException(reason + "of type: " + Util.typeName(pojo));
+    // Convert POJOs
+    Object converted = CustomClassMapper.convertToPlainJavaTypes(input);
+
+    if (source == UserData.Source.Set && !(converted instanceof Map)) {
+      throw new IllegalArgumentException(badDocReason + "of type: " + Util.typeName(input));
     }
 
-    @SuppressWarnings("unchecked") // CustomClassMapper promises to map keys to Strings.
-    Map<String, Object> map = (Map<String, Object>) converted;
-    return map;
+    // Parse to Model types.
+    return parseData(converted, context);
   }
 
   /**
-   * Internal helper for parsing user data.
+   * Recursive helper for parsing user data.
    *
    * @param input Data to be parsed.
    * @param context A context object representing the current path being parsed, the source of the
@@ -416,7 +416,7 @@ public final class UserDataConverter {
       // being unioned or removed are not considered writes since they cannot
       // contain any FieldValue sentinels, etc.
       ParseContext context = accumulator.rootContext();
-      result.add(parseData(element, context.childContext(i)));
+      result.add(convertAndParseData(element, context.childContext(i)));
     }
     return result;
   }