+++ firebase-firestore/src/main/java/com/google/firebase/firestore/WriteBatch.java
@@ -59,12 +59,13 @@ public class WriteBatch {
    * yet exist, it will be created. If a document already exists, it will be overwritten.
    *
    * @param documentRef The DocumentReference to overwrite.
-   * @param data A map of the fields and values for the document.
+   * @param data The data to write to the document (e.g. a Map or a POJO containing the desired
+   *     document contents).
    * @return This WriteBatch instance. Used for chaining method calls.
    */
   @NonNull
   @PublicApi
-  public WriteBatch set(@NonNull DocumentReference documentRef, @NonNull Map<String, Object> data) {
+  public WriteBatch set(@NonNull DocumentReference documentRef, @NonNull Object data) {
     return set(documentRef, data, SetOptions.OVERWRITE);
   }
 
@@ -74,18 +75,18 @@ public class WriteBatch {
    * into an existing document.
    *
    * @param documentRef The DocumentReference to overwrite.
-   * @param data A map of the fields and values for the document.
+   * @param data The data to write to the document (e.g. a Map or a POJO containing the desired
+   *     document contents).
    * @param options An object to configure the set behavior.
    * @return This WriteBatch instance. Used for chaining method calls.
    */
   @NonNull
   @PublicApi
   public WriteBatch set(
-      @NonNull DocumentReference documentRef,
-      @NonNull Map<String, Object> data,
-      @NonNull SetOptions options) {
+      @NonNull DocumentReference documentRef, @NonNull Object data, @NonNull SetOptions options) {
     firestore.validateReference(documentRef);
     checkNotNull(data, "Provided data must not be null.");
+    checkNotNull(options, "Provided options must not be null.");
     verifyNotCommitted();
     ParsedSetData parsed =
         options.isMerge()
@@ -93,37 +94,6 @@ public class WriteBatch {
             : firestore.getDataConverter().parseSetData(data);
     mutations.addAll(parsed.toMutationList(documentRef.getKey(), Precondition.NONE));
     return this;
-  }
-
-  /**
-   * Overwrites the document referred to by the provided DocumentReference. If the document does not
-   * yet exist, it will be created. If a document already exists, it will be overwritten.
-   *
-   * @param documentRef The DocumentReference to overwrite.
-   * @param pojo The POJO that will be used to populate the document contents.
-   * @return This WriteBatch instance. Used for chaining method calls.
-   */
-  @NonNull
-  @PublicApi
-  public WriteBatch set(@NonNull DocumentReference documentRef, @NonNull Object pojo) {
-    return set(documentRef, firestore.getDataConverter().convertPOJO(pojo), SetOptions.OVERWRITE);
-  }
-
-  /**
-   * Writes to the document referred to by the provided DocumentReference. If the document does not
-   * yet exist, it will be created. If you pass {@link SetOptions}, the provided data can be merged
-   * into an existing document.
-   *
-   * @param documentRef The DocumentReference to overwrite.
-   * @param pojo The POJO that will be used to populate the document contents.
-   * @param options An object to configure the set behavior.
-   * @return This WriteBatch instance. Used for chaining method calls.
-   */
-  @NonNull
-  @PublicApi
-  public WriteBatch set(
-      @NonNull DocumentReference documentRef, @NonNull Object pojo, @NonNull SetOptions options) {
-    return set(documentRef, firestore.getDataConverter().convertPOJO(pojo), options);
   }
 
   /**